backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.8.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
passlib[argon2]==1.7.4
PyJWT==2.9.0
redis==5.0.7
celery==5.4.0
requests==2.32.3
boto3==1.35.0
python-dotenv==1.0.1
tensorflow-cpu==2.16.2
transformers==4.44.2
torch==2.4.0
numpy==1.26.4
pandas==2.2.2
